backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
Werkzeug==3.1.3
gunicorn==23.0.0
python-dotenv==1.0.1
